Remove requirement for settings.py and fix s3 bug
Allow the user to use a default directory of site_path/output to
compile their site.

A bug in the s3 uploader's dup checker would cause a thread to die and
the parent to lock.  We weren't checking if we succesfully got the
s3_key before asking for its md5.  Problem fixed.